Refactored throw new assertion error
