+++ src/test/java/seedu/address/logic/commands/AddCommandTest.java
@@ -20,6 +20,7 @@ import javafx.collections.ObservableList;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.logic.Messages;
 import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.logic.commands.exceptions.InappropriateMethodCallAssertionError;
 import seedu.address.model.AddressBook;
 import seedu.address.model.CourseName;
 import seedu.address.model.Model;
